+++ src/data/SchedulesDB.java
@@ -1,5 +1,7 @@
 package data;
 
+import iot.Manager;
+
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -39,6 +41,7 @@ public class SchedulesDB extends NoSQLDB<Schedule, List<Schedule>> {
 
     @Override
     public void correctInitialState() {
+        if (Manager.MODE != 2) return;
         dbClient.context().deleteDB("schedules", "delete database");
         dbClient.context().createDB("schedules");
     }
